requirements.txt
python-dotenv==1.0.0requests==2.31.0
